version final

+++ src/metrics.py
@@ -8,7 +8,7 @@ def mse(y_true, y_pred):
     y_pred = y_pred.reshape(-1, 1)
 
     err = y_pred - y_true
-    res = np.mean(err**2)
+    res = float(np.mean(err**2))
     
     return res
 
@@ -18,9 +18,9 @@ def mae(y_true, y_pred):
     y_pred = y_pred.reshape(-1, 1)
 
     err = y_pred - y_true
-    res = np.mean(np.abs(err))
+    res = float(np.mean(np.abs(err)))
 
     return res
 
 def rmse(y_true, y_pred):
-    return np.sqrt(mse(y_true, y_pred))
+    return float(np.sqrt(mse(y_true, y_pred)))
